Extract log getting logic to its own hook remove console.logs
