+++ packages/nextjs/hooks/scaffold-eth/index.ts
@@ -14,3 +14,4 @@ export * from "./useScaffoldEventHistory";
 export * from "./useTransactor";
 export * from "./useFetchBlocks";
 export * from "./useSearchHandler";
+export * from "./useContractLogs";
